+++ lang/it/screen/permissions.php
@@ -1,7 +1,6 @@
 <?php
 
 return [
-    'access' => 'Permission to access the :screen screen.',
     'hola_contexto' => [
         'access' => 'Permission to access Hola Contexto.',
     ],